Revert "[Draft] Introduce generic getMigrateDashboards function"
This reverts commit ff031ad8bbb09a2c9b72ce12f5f91d6a8201e410.
